Add Apache 2.0 license
This copies the Apache 2.0 license from SPIRV-Cross. Since crossc is
just a thin wrapper for SPIRV-Cross, using the same license is the
simplest option.

Also yeah, I forgot to add license information to the previous releases
of crossc. They may also be used under the terms of the Apache 2.0
license.